elasticsearch/requirements.txt
asttokens==3.0.0
certifi==2025.8.3
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
elastic-transport==9.1.0
elasticsearch==9.1.0
executing==2.2.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyzmq==27.0.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.14
wcwidth==0.2.13
